refactor(api): move assessment-result service
