+++ api/lib/application/certification-courses/certification-course-controller.js
@@ -8,7 +8,7 @@ import { usecases } from '../../domain/usecases/index.js';
 import { DomainTransaction } from '../../infrastructure/DomainTransaction.js';
 import { AssessmentResult } from '../../../src/shared/domain/models/AssessmentResult.js';
 import { CompetenceMark } from '../../domain/models/CompetenceMark.js';
-import * as assessmentResultService from '../../domain/services/assessment-result-service.js';
+import * as assessmentResultService from '../../../src/shared/domain/services/assessment-result-service.js';
 
 import { extractLocaleFromRequest } from '../../infrastructure/utils/request-response-utils.js';
