+++ api/src/shared/domain/services/assessment-result-service.js
@@ -1,6 +1,6 @@
-import * as assessmentResultRepository from '../../infrastructure/repositories/assessment-result-repository.js';
-import * as competenceMarkRepository from '../../infrastructure/repositories/competence-mark-repository.js';
-import { CompetenceMark } from '../models/CompetenceMark.js';
+import * as assessmentResultRepository from '../../../../lib/infrastructure/repositories/assessment-result-repository.js';
+import * as competenceMarkRepository from '../../../../lib/infrastructure/repositories/competence-mark-repository.js';
+import { CompetenceMark } from '../../../../lib/domain/models/CompetenceMark.js';
 import bluebird from 'bluebird';
 
 async function _validatedDataForAllCompetenceMark(competenceMarks) {
